refactor(view, error): 파일 분리 및 정리
