+++ src/view/inputBonusNumber.js
