+++ src/view/inputMoney.js
